Reorg buildrunner tests distinuguished by environment
