+++ test-data/integration/ibm/buildrunner/docker/docker-hf/space/assetstores/hf/store.yaml
