--- test-data/integration/ibm/buildwatcher/1step/skip/build.yaml
@@ -1,35 +1,0 @@
-llm.build:
-  name: skip_test_2step
-  targets:
-    download_file1:
-      environment_uri: space://environments/{{ space.variables.DEFAULT_CPU_ENVIRONMENT }}
-      inputs:
-        seed_data:
-          uri: lh://staging/granite_dot_build.public/tables/maximo_digit_input1
-      outputs:
-        download_file1:
-          uri: lh://staging/granite_dot_build.public/tables/test_skip_step_data_{{ binding.path | short_hash }}
-      steps:
-        - step_uri: space://steps/download
-          config:
-            download_config:
-              url1: https://huggingface.co/datasets/datasets-examples/doc-formats-jsonl-1/resolve/main/data.jsonl
-              url2: https://huggingface.co/datasets/datasets-examples/doc-formats-jsonl-1/resolve/main/data.jsonl
-              url3: https://huggingface.co/datasets/datasets-examples/doc-formats-jsonl-1/resolve/main/data.jsonl
-              output_dir: /gb-read-write/downloads_{{ run_metadata.targetsteprun_id | short_hash }}
-    download_file2:
-      environment_uri: space://environments/{{ space.variables.DEFAULT_CPU_ENVIRONMENT }}
-      inputs:
-        seed_data:
-          binding: download_file1.download_file1
-      outputs:
-        download_file2:
-          uri: lh://staging/granite_dot_build.public/tables/test_skip_process_data2_{{ binding.path | short_hash }}
-      steps:
-        - step_uri: space://steps/download
-          config:
-            download_config:
-              url1: https://huggingface.co/datasets/datasets-examples/doc-formats-jsonl-1/resolve/main/data.jsonl
-              url2: https://huggingface.co/datasets/datasets-examples/doc-formats-jsonl-1/resolve/main/data.jsonl
-              url3: https://huggingface.co/datasets/datasets-examples/doc-formats-jsonl-1/resolve/main/data.jsonl
-              output_dir: /gb-read-write/process_{{ run_metadata.targetsteprun_id | short_hash }}
